Update <Home> page component to use <Layout>
