+++ src/ui/pages/Home/Home.test.tsx
@@ -1,12 +1,12 @@
 import * as React from "react";
 import { render } from "@testing-library/react";
 
-import { HomeView } from "./";
+import { HomeUnwrapped } from "./";
 
 describe("Home", () => {
   it("Renders a header", () => {
     const result = render(
-      <HomeView
+      <HomeUnwrapped
         counter={0}
         incrementCounterNow={() => null}
         incrementCounterInNSeconds={() => null}
