--- src/ui/pages/Home/index.less
@@ -1,8 +1,0 @@
-@import (reference) "~clever-components/dist/less/index";
-
-// TODO: flesh out true homepage styles
-
-.pages--Home > h1 {
-  background: @primary_blue;
-  color: @neutral_white;
-}
